框架整合第一版 能访问页面

+++ boke-mapper/src/main/java/com/lsh/mapper/boke/BokeTopMapper.java
@@ -5,5 +5,5 @@ import com.pearadmin.boke.entry.BokeTop;
 
 public interface BokeTopMapper extends BaseMapper<BokeTop> {
     
-    int getByuserIdCount(Integer userId);
+    int getByuserIdCount(Long userId);
 }
