+++ boke-mapper/src/main/java/com/lsh/mapper/boke/BokesMapper.java
@@ -16,11 +16,11 @@ import com.pearadmin.boke.vo.BokeListEntry;
 import com.pearadmin.boke.vo.query.QueryBokeVo;
 
 public interface BokesMapper extends BaseMapper<Bokes> {
-    IPage<BokeListEntry> getBokes(IPage<BokeListEntry> ipage, String title,Integer labelId,Integer cateId, String createDate, Integer userId);
+    IPage<BokeListEntry> getBokes(IPage<BokeListEntry> ipage, String title,Integer labelId,Integer cateId, String createDate, Long userId);
     IPage<BokeListEntry> getBokesList(IPage ipage,@Param("vo") QueryBokeVo vo);
     
-    BokeListEntry getBokesById(Integer bokeId,Integer userId);
-    BokeListEntry getBokesByEmId(Integer bokeId,Integer userId);
+    BokeListEntry getBokesById(Integer bokeId,Long userId);
+    BokeListEntry getBokesByEmId(Integer bokeId,Long userId);
 
     List<CategoryCount> getCateCount(Integer userId);
