+++ pear-modules/pear-system/src/main/java/com/pearadmin/boke/service/impl/BokeTopServiceImpl.java
@@ -1,15 +1,15 @@
 package com.pearadmin.boke.service.impl;
 
-import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
-import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
-import com.pearadmin.boke.entry.BokeTop;
-import com.lsh.mapper.boke.BokeTopMapper;
-import com.pearadmin.boke.service.BokeTopService;
+import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
+import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
+import com.lsh.mapper.boke.BokeTopMapper;
+import com.pearadmin.boke.entry.BokeTop;
+import com.pearadmin.boke.service.BokeTopService;
 
 @Service
 public class BokeTopServiceImpl extends ServiceImpl<BokeTopMapper, BokeTop> implements BokeTopService {
@@ -18,7 +18,7 @@ public class BokeTopServiceImpl extends ServiceImpl<BokeTopMapper, BokeTop> impl
     private BokeTopMapper bokeTopMapper;
     
     @Override
-    public List<BokeTop> getByuserId(Integer userId) {
+    public List<BokeTop> getByuserId(Long userId) {
         QueryWrapper<BokeTop> wrapper = new QueryWrapper<>();
         wrapper.eq("user_id",userId);
         List<BokeTop> bokeTops = bokeTopMapper.selectList(wrapper);
@@ -26,7 +26,7 @@ public class BokeTopServiceImpl extends ServiceImpl<BokeTopMapper, BokeTop> impl
     }
 
     @Override
-    public int getByuserIdCount(Integer userId) {
+    public int getByuserIdCount(Long userId) {
         return bokeTopMapper.getByuserIdCount(userId);
     }
 }
